refactor landing component file structure
